+++ bin/install-wp-tests.sh
@@ -254,7 +254,7 @@ install_activitypub_plugin() {
 		return;
 	fi
 
-	PLUGIN_VERSION="trunk"
+	PLUGIN_VERSION="5324992"
 
     URL="https://github.com/Automattic/wordpress-activitypub"
 
@@ -270,13 +270,13 @@ install_wp_plugins() {
 	# Install the one and only ActivityPub plugin (greetings @pfefferle).
 	install_activitypub_plugin
 	# Install (not-activate) all supported event plugins.
-	install_wp_plugin the-events-calendar "6.8.1"
+	install_wp_plugin the-events-calendar "6.10.0"
 	install_wp_plugin very-simple-event-list
 	install_wp_plugin gatherpress
 	install_wp_plugin eventprime-event-calendar-management
 	install_wp_plugin events-manager "6.6.3"
 	install_wp_plugin wp-event-manager "3.1.45.1"
-	install_wp_plugin wp-event-solution "4.0.14"
+	install_wp_plugin wp-event-solution "4.0.20"
 	install_wp_plugin event-organiser "3.12.8"
 	# Mec is not installable via wordpress.org, we use our own mirror.
 	install_wp_plugin_mec