+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Changed
+
+* Changed: Now depends on ActivityPub plugin version greater than 5.0.0
+*
 ### Added
 
 * Event self-announce feature at configurable time before event starts